Updated CSS/Created BandProfile.jsx to export component to index for rendering/File clean-up
