+++ package.json
@@ -27,6 +27,7 @@
     "chai-http": "^4.3.0",
     "chalk": "^4.1.0",
     "express": "^4.17.1",
+    "favicons": "^6.2.0",
     "jquery": "^3.5.1",
     "mdbreact": "^4.27.0",
     "mongoose": "^5.10.6",
@@ -44,9 +45,11 @@
     "@babel/preset-env": "^7.11.5",
     "@babel/preset-react": "^7.10.4",
     "babel-loader": "^8.1.0",
+    "css-loader": "^4.3.0",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
     "jest": "^26.4.2",
     "mocha": "^8.1.3",
+    "style-loader": "^2.0.0",
     "supertest": "^5.0.0",
     "webpack": "^4.44.2",
     "webpack-cli": "^3.3.12"
